refactor: migrating rest of the sources to v2
