+++ src/controllers/util/index.test.ts
@@ -32,13 +32,20 @@ describe('adaptInputToVersion', () => {
     expect(result).toEqual(expected);
   });
   it('should return the input unchanged when the implementation version and request version are the same i.e. v0', () => {
-    const sourceType = 'pipedream';
+    jest
+      .spyOn(ControllerUtility as any, 'getSourceVersionsMap')
+      .mockReturnValue(new Map([['someSourceType', 'v0']]));
+
+    const sourceType = 'someSourceType';
     const requestVersion = 'v0';
     const input = [
       { key1: 'val1', key2: 'val2' },
       { key1: 'val1', key2: 'val2' },
       { key1: 'val1', key2: 'val2' },
     ];
+
+    // Mock return value for getSourceVersionsMap
+
     const expected = {
       implementationVersion: 'v0',
       input: [
@@ -52,8 +59,13 @@ describe('adaptInputToVersion', () => {
 
     expect(result).toEqual(expected);
   });
+
   it('should return the input unchanged when the implementation version and request version are the same i.e. v1', () => {
-    const sourceType = 'webhook';
+    jest
+      .spyOn(ControllerUtility as any, 'getSourceVersionsMap')
+      .mockReturnValue(new Map([['someSourceType', 'v1']]));
+
+    const sourceType = 'someSourceType';
     const requestVersion = 'v1';
     const input = [
       {
@@ -69,6 +81,7 @@ describe('adaptInputToVersion', () => {
         source: { id: 'source_id', config: { configField1: 'configVal1' } },
       },
     ];
+
     const expected = {
       implementationVersion: 'v1',
       input: [
@@ -97,8 +110,13 @@ describe('adaptInputToVersion', () => {
 
     expect(result).toEqual(expected);
   });
+
   it('should convert input from v0 to v1 when the request version is v0 and the implementation version is v1', () => {
-    const sourceType = 'webhook';
+    jest
+      .spyOn(ControllerUtility as any, 'getSourceVersionsMap')
+      .mockReturnValue(new Map([['someSourceType', 'v1']]));
+
+    const sourceType = 'someSourceType';
     const requestVersion = 'v0';
     const input = [
       { key1: 'val1', key2: 'val2' },
@@ -120,7 +138,11 @@ describe('adaptInputToVersion', () => {
   });
 
   it('should convert input from v1 to v0 format when the request version is v1 and the implementation version is v0', () => {
-    const sourceType = 'pipedream';
+    jest
+      .spyOn(ControllerUtility as any, 'getSourceVersionsMap')
+      .mockReturnValue(new Map([['someSourceType', 'v0']]));
+
+    const sourceType = 'someSourceType';
     const requestVersion = 'v1';
     const input = [
       {
@@ -151,7 +173,11 @@ describe('adaptInputToVersion', () => {
   });
 
   it('should convert input from v2 to v0 format when the request version is v2 and the implementation version is v0', () => {
-    const sourceType = 'pipedream';
+    jest
+      .spyOn(ControllerUtility as any, 'getSourceVersionsMap')
+      .mockReturnValue(new Map([['someSourceType', 'v0']]));
+
+    const sourceType = 'someSourceType';
     const requestVersion = 'v2';
 
     const input = [
@@ -204,7 +230,11 @@ describe('adaptInputToVersion', () => {
   });
 
   it('should fail trying to convert input from v2 to v0 format when the request version is v2 and the implementation version is v0', () => {
-    const sourceType = 'pipedream';
+    jest
+      .spyOn(ControllerUtility as any, 'getSourceVersionsMap')
+      .mockReturnValue(new Map([['someSourceType', 'v0']]));
+
+    const sourceType = 'someSourceType';
     const requestVersion = 'v2';
 
     const input = [
@@ -235,7 +265,11 @@ describe('adaptInputToVersion', () => {
   });
 
   it('should convert input from v2 to v1 format when the request version is v2 and the implementation version is v1', () => {
-    const sourceType = 'webhook';
+    jest
+      .spyOn(ControllerUtility as any, 'getSourceVersionsMap')
+      .mockReturnValue(new Map([['someSourceType', 'v1']]));
+
+    const sourceType = 'someSourceType';
     const requestVersion = 'v2';
 
     const input = [
@@ -303,7 +337,11 @@ describe('adaptInputToVersion', () => {
   });
 
   it('should fail trying to convert input from v2 to v1 format when the request version is v2 and the implementation version is v1', () => {
-    const sourceType = 'webhook';
+    jest
+      .spyOn(ControllerUtility as any, 'getSourceVersionsMap')
+      .mockReturnValue(new Map([['someSourceType', 'v1']]));
+
+    const sourceType = 'someSourceType';
     const requestVersion = 'v2';
 
     const input = [
@@ -335,7 +373,11 @@ describe('adaptInputToVersion', () => {
 
   // Should return an empty array when the input is an empty array
   it('should return an empty array when the input is an empty array', () => {
-    const sourceType = 'pipedream';
+    jest
+      .spyOn(ControllerUtility as any, 'getSourceVersionsMap')
+      .mockReturnValue(new Map([['someSourceType', 'v0']]));
+
+    const sourceType = 'someSourceType';
     const requestVersion = 'v1';
     const input = [];
     const expected = { implementationVersion: 'v0', input: [] };
@@ -349,7 +391,6 @@ describe('adaptInputToVersion', () => {
     const sourceType = 'someSourceType';
     const requestVersion = 'v1';
 
-    // Mock return value for getSourceVersionsMap
     jest
       .spyOn(ControllerUtility as any, 'getSourceVersionsMap')
       .mockReturnValue(new Map([['someSourceType', 'v2']]));
