+++ src/sources/signl4/mapping.json
