+++ src/sources/slack/mapping.json
