+++ src/sources/slack/util.js
