--- src/v0/sources/segment/transform.js
@@ -1,5 +1,0 @@
-function process(events) {
-  return events;
-}
-
-exports.process = process;
