+++ src/v1/sources/shopify/webpixelTransformations/pixelUtils.js
@@ -1,6 +1,6 @@
 /* eslint-disable no-param-reassign */
 const { isDefinedAndNotNull } = require('@rudderstack/integrations-lib');
-const Message = require('../../../../v0/sources/message');
+const Message = require('../../../../sources/message');
 const { EventType } = require('../../../../constants');
 const {
   INTEGERATION,
